# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1",
    "title": "Your pet API",
    "description": "This is API for selling pets, adding them to favorites, describing your pet, etc.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [
    { "url": "https://your-pet-shw3.onrender.com" },
    { "url": "http://localhost:5000" }
  ],
...
  "paths": {
...
    "/notices": {
      "get": {
        "tags": ["Notice"],
        "summary": "Get notices by category and/or search parameters",
        "parameters": [
          {
            "in": "query",
            "name": "NoticesCategoriesNav",
            "description": "Notices category",
            "explode": true,
            "schema": {
              "type": "string",
              "default": "sell",
              "enum": ["sell", "lost-found", "in-good-hands"]
            }
          },
          {
            "in": "query",
            "name": "NoticesSearch",
            "description": "Notices search parameter (search by title)",
            "schema": {
              "type": "string",
              "default": "empty string",
              "example": "cat"
            }
          },
          {
            "in": "query",
            "name": "page",
            "description": "Current page notices (page=allNotices/limit)",
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1,
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "The number of notices that should be on one page",
            "schema": {
              "type": "number",
              "default": 12,
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNoticesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorServerResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Notice"],
        "summary": "Add notices by chose category",
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "description": "Notice's object (multipart/form-data)",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddNoticeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorServerResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/notices/{noticeId}": {
      "get": {
        "tags": ["Notice"],
        "summary": "Get notice by id",
        "parameters": [
          {
            "in": "path",
            "name": "noticeId",
            "required": true,
            "description": "Notice's id",
            "schema": {
              "type": "string",
              "example": "64e251e974992f975f71c4ee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNoticeByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorServerResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Notice"],
        "summary": "Delete owner notice by id",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "noticeId",
            "required": true,
            "description": "Notice's id",
            "schema": {
              "type": "string",
              "example": "64e251e974992f975f71c4ee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveOwnerNoticeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorServerResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetNoticesResponse": {
        "type": "object",
        "properties": {
          "notices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Notice's title",
                  "example": "Cute cat looking for a home"
                },
                "petAvatarURL": {
                  "type": "string",
                  "description": "Pet's foto link",
                  "example": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg"
                },
                "name": {
                  "type": "string",
                  "description": "Pet's name",
                  "example": "Melany"
                },
                "birthday": {
                  "type": "string",
                  "description": "Pet's birthday",
                  "example": "21-09-2023"
                },
                "type": {
                  "type": "string",
                  "description": "Pet's type",
                  "example": "Promeranian"
                },
                "sex": {
                  "type": "string",
                  "description": "Pet's sex",
                  "enum": ["male", "female"],
                  "example": "male"
                },
                "comments": {
                  "type": "string",
                  "description": "Notice's comments",
                  "example": "Rich would be the perfect addition to an active family that loves to play."
                },
                "location": {
                  "type": "string",
                  "description": "Pet's location",
                  "example": "Barnstable Town"
                },
                "category": {
                  "type": "string",
                  "description": "Notice's category",
                  "enum": ["sell", "lost/found", "in good hands"],
                  "example": "lost/found"
                },
                "price": {
                  "type": "number",
                  "description": "Pet's price",
                  "example": "23"
                },
                "owner": {
                  "type": "string",
                  "description": "Owner's id",
                  "example": "534d677fcc8ea74aedaabdc1"
                },
                "idUsersAddedFavorite": {
                  "type": "array",
                  "description": "User's id who added notice favorite",
                  "item": {
                    "type": "string",
                    "describe": "User's id",
                    "example": "64df88ca1f2426eddf5014d7"
                  }
                }
              }
            }
          },
          "totalPages": {
            "type": "number",
            "description": "The number of total pages",
            "example": 1
          }
        },
        "example": {
          "notices": [
            {
              "title": "Cute pet looking for a home",
              "petAvatarURL": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg",
              "name": "Charlie",
              "birthday": "21-09-2023",
              "type": "Promeranian",
              "sex": "male",
              "place": "Little Rock",
              "comments": "Rich would be the perfect addition to an active family that loves to play.",
              "category": "sell",
              "price": "40",
              "owner": "f24cf26ff2b7a80e999b8abb",
              "idUsersAddedFavorite": [
                "64df88ca1f2426eddf5014d7",
                "75df88ca1f6346eddf5014g6"
              ]
            },
            {
              "title": "Cute pet looking for a home",
              "petAvatarURL": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg",
              "name": "Charlie",
              "birthday": "21-09-2023",
              "type": "Promeranian",
              "sex": "male",
              "place": "Little Rock",
              "comments": "Rich would be the perfect addition to an active family that loves to play.",
              "category": "sell",
              "price": "40",
              "owner": "f24cf26ff2b7a80e999b8abb",
              "idUsersAddedFavorite": [
                "64df88ca1f2426eddf5014d7",
                "75df88ca1f6346eddf5014g6"
              ]
            }
          ],
          "totalPages": 3
        }
      },
...
      "GetNoticeByIdResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Notice's title",
            "example": "Cute cat looking for a home"
          },
          "petAvatarURL": {
            "type": "string",
            "description": "Pet's foto link",
            "example": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "example": "Melany"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "example": "21-09-2023"
          },
          "type": {
            "type": "string",
            "description": "Pet's type",
            "example": "Promeranian"
          },
          "sex": {
            "type": "string",
            "description": "Pet's sex",
            "enum": ["male", "female"],
            "example": "male"
          },
          "comments": {
            "type": "string",
            "description": "Notice's comments",
            "example": "Rich would be the perfect addition to an active family that loves to play."
          },
          "location": {
            "type": "string",
            "description": "Pet's location",
            "example": "Barnstable Town"
          },
          "category": {
            "type": "string",
            "description": "Notice's category",
            "enum": ["sell", "lost/found", "in good hands"],
            "example": "lost/found"
          },
          "price": {
            "type": "number",
            "description": "Pet's price",
            "example": "23"
          },
          "owner": {
            "type": "string",
            "description": "Owner's id",
            "example": "534d677fcc8ea74aedaabdc1"
          },
          "idUsersAddedFavorite": {
            "type": "array",
            "description": "User's id who added notice favorite",
            "item": {
              "type": "string",
              "describe": "User's id",
              "example": "64df88ca1f2426eddf5014d7"
            }
          }
        },
        "example": {
          "title": "Cute pet looking for a home",
          "petAvatarURL": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg",
          "name": "Charlie",
          "birthday": "21-09-2023",
          "type": "Promeranian",
          "sex": "male",
          "place": "Little Rock",
          "comments": "Rich would be the perfect addition to an active family that loves to play.",
          "category": "sell",
          "price": "40",
          "owner": "f24cf26ff2b7a80e999b8abb",
          "idUsersAddedFavorite": [
            "64df88ca1f2426eddf5014d7",
            "75df88ca1f6346eddf5014g6"
          ]
        }
      },
...
      "AddNoticeRequest": {
        "type": "object",
        "required": [
          "title",
          "petAvatar",
          "name",
          "birthday",
          "location",
          "type",
          "sex",
          "category",
          "price"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Notice's title",
            "example": "Cute cat looking for a home"
          },
          "petAvatar": {
            "type": "file",
            "format": "binary",
            "description": "Pet's avatar file",
            "example": "cat.jpg"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "example": "Melany"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "example": "21-09-2023"
          },
          "type": {
            "type": "string",
            "description": "Pet's type",
            "example": "Promeranian"
          },
          "sex": {
            "type": "string",
            "description": "Pet's sex",
            "enum": ["male", "female"],
            "example": "male"
          },
          "comments": {
            "type": "string",
            "description": "Notice's comments",
            "example": "Rich would be the perfect addition to an active family that loves to play."
          },
          "location": {
            "type": "string",
            "description": "Pet's location",
            "example": "Barnstable Town"
          },
          "category": {
            "type": "string",
            "description": "Notice's category",
            "enum": ["sell", "lost/found", "in good hands"],
            "example": "lost/found"
          },
          "price": {
            "type": "number",
            "description": "Pet's price",
            "example": "23"
          }
        }
      },
      "AddNoticeResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Notice's title",
            "example": "Cute cat looking for a home"
          },
          "petAvatarURL": {
            "type": "string",
            "description": "Pet's foto link",
            "example": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "example": "Melany"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "example": "21-09-2023"
          },
          "type": {
            "type": "string",
            "description": "Pet's type",
            "example": "Promeranian"
          },
          "sex": {
            "type": "string",
            "description": "Pet's sex",
            "enum": ["male", "female"],
            "example": "male"
          },
          "comments": {
            "type": "string",
            "description": "Notice's comments",
            "example": "Rich would be the perfect addition to an active family that loves to play."
          },
          "location": {
            "type": "string",
            "description": "Pet's location",
            "example": "Barnstable Town"
          },
          "category": {
            "type": "string",
            "description": "Notice's category",
            "enum": ["sell", "lost/found", "in good hands"],
            "example": "lost/found"
          },
          "price": {
            "type": "number",
            "description": "Pet's price",
            "example": "23"
          },
          "owner": {
            "type": "string",
            "description": "Owner's id",
            "example": "534d677fcc8ea74aedaabdc1"
          },
          "idUsersAddedFavorite": {
            "type": "array",
            "description": "User's id who added notice favorite",
            "item": {
              "type": "string",
              "describe": "User's id",
              "example": "64df88ca1f2426eddf5014d7"
            }
          }
        },
        "example": {
          "title": "Cute pet looking for a home",
          "petAvatarURL": "https://res.cloudinary.com/dozyx4svd/image/upload/v1693047085/pets/cujdax5sgppc1g9mikdb.jpg",
          "name": "Charlie",
          "birthday": "21-09-2023",
          "type": "Promeranian",
          "sex": "male",
          "place": "Little Rock",
          "comments": "Rich would be the perfect addition to an active family that loves to play.",
          "category": "sell",
          "price": "40",
          "owner": "f24cf26ff2b7a80e999b8abb",
          "idUsersAddedFavorite": [
            "64df88ca1f2426eddf5014d7",
            "75df88ca1f6346eddf5014g6"
          ]
        }
      },
...
      "RemoveOwnerNoticeResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Successful operation",
            "example": "Notice successfully delete"
          }
        }
      },
...
      "ErrorUnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Unauthorized"
          }
        }
      },
      "ErrorServerResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "This text will describe what is wrong with the internal server"
          }
        }
      },
      "ErrorNotFoundResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Not Found"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}